--- op_tests/triton_tests/utils/accuracy_analysis.py
@@ -1,29 +1,0 @@
-import torch
-
-
-def compare_accuracy(current, reference):
-    """Print quick statistics comparing FP8 and SageAttn tensors."""
-    current_f = current.float()
-    reference_f = reference.float()
-    abs_diff = torch.abs(reference_f - current_f)
-
-    print("Output Tensor Stats:")
-    print(
-        f"  Reference ({tuple(reference_f.shape)}): min={reference_f.min().item():.6f}, max={reference_f.max().item():.6f}, "
-        f"mean={reference_f.mean().item():.6f}, std={reference_f.std().item():.6f}"
-    )
-    print(
-        f"  Test      ({tuple(current_f.shape)}): min={current_f.min().item():.6f}, max={current_f.max().item():.6f}, "
-        f"mean={current_f.mean().item():.6f}, std={current_f.std().item():.6f}"
-    )
-
-    print("Correctness Comparison:")
-    print(f"  Mean Absolute Error: {abs_diff.mean().item():.6e}")
-    print(f"  Max Absolute Error: {abs_diff.max().item():.6e}")
-    print(f"  Std Absolute Error: {abs_diff.std().item():.6e}")
-    ref_flat = reference_f.reshape(-1)
-    test_flat = current_f.reshape(-1)
-    cos_sim = torch.nn.functional.cosine_similarity(
-        ref_flat.unsqueeze(0), test_flat.unsqueeze(0)
-    )
-    print(f"  Cosine Similarity: {cos_sim.item():.8f}")
